Fixed Issue#6: this is a very big and frightening change.
The file service.py was filled with many static classes (the core of journeytag), and I moved them into seperate modules (the python way to do things).

No functionality changed during all this, it is purely code restructuring.

JTTestHarness passes 100%, but I am worried about bugs that it doesn't test for. I can check these by a few hours of using the app locally.  Yuck!!

Lastly, the bugs introduced are going to be caused by referencing old modules and fixing this is the easy part.