Related to Issue#6
A slight improvement to the method of responding to errors.  Reciving back non-JSON from the server will always result in a "didFail" response.

Essentally, this stops my test suite from running so I can read the error.

Previous to this change, the test runner would keep going, making it impossible to read log output.

This may be useful to merge into the JourneyTagClient.